requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
python-dotenv==1.0.1
Werkzeug==3.1.3
PyPDF2==3.0.1
docx==0.2.4
google-generativeai==0.3.1
